+++ deployables/app/e2e/accountHandling/lockedAccount.spec.ts
@@ -1,45 +1,10 @@
 import test, { expect } from '@playwright/test'
-import { Chain } from '@zodiac/chains'
 import { encode } from '@zodiac/schema'
-import {
-  createMockEndWaypoint,
-  createMockEoaAccount,
-  createMockExecutionRoute,
-  createMockOwnsConnection,
-  createMockStartingWaypoint,
-  createMockTransaction,
-  createMockWaypoints,
-} from '@zodiac/test-utils'
 import { mockWeb3 } from '@zodiac/test-utils/e2e'
-import { prefixAddress, unprefixAddress } from 'ser-kit'
 import { connectWallet } from '../connectWallet'
+import { account, route, transaction } from './fixture'
 
 test.describe('Locked account', () => {
-  const account = '0xf06a3a90f9a248630bd15f9debc7a6a20e54677d'
-
-  const initiator = prefixAddress(undefined, account)
-
-  const route = createMockExecutionRoute({
-    initiator,
-    avatar: prefixAddress(
-      Chain.GNO,
-      '0xb1578ecfa1da7405821095aa3612158926e6a72a',
-    ),
-    waypoints: createMockWaypoints({
-      start: createMockStartingWaypoint(
-        createMockEoaAccount({ address: unprefixAddress(initiator) }),
-      ),
-      end: createMockEndWaypoint({
-        account: {
-          chainId: Chain.GNO,
-          address: '0xb1578ecfa1da7405821095aa3612158926e6a72a',
-        },
-        connection: createMockOwnsConnection(initiator),
-      }),
-    }),
-  })
-  const transaction = createMockTransaction()
-
   test('handles wallet disconnect gracefully', async ({ page }) => {
     const { lockWallet } = await mockWeb3(page, {
       accounts: [account],
