+++ deployables/app/e2e/accountHandling/wrongChain.spec.ts
@@ -1,15 +1,21 @@
 import test, { expect } from '@playwright/test'
+import { Chain } from '@zodiac/chains'
+import { encode } from '@zodiac/schema'
 import { mockWeb3 } from '@zodiac/test-utils/e2e'
 import { connectWallet } from '../connectWallet'
+import { account, route, transaction } from './fixture'
 
 test.describe('Wrong chain selected', () => {
   test('it is possible to switch to the correct chain', async ({ page }) => {
-    const { switchChain } = await mockWeb3(page)
+    const { switchChain } = await mockWeb3(page, {
+      accounts: [account],
+      chain: 'gnosis',
+    })
 
-    await page.goto('/new-route')
+    await page.goto(`/submit/${encode(route)}/${encode([transaction])}`)
 
-    await connectWallet(page)
-    await switchChain(10)
+    await connectWallet(page, account)
+    await switchChain(Chain.OETH)
 
     await expect(
       page.getByRole('alert', { name: 'Chain mismatch' }),
@@ -19,15 +25,18 @@ test.describe('Wrong chain selected', () => {
   test('it is possible to switch back to the connected chain', async ({
     page,
   }) => {
-    const { switchChain } = await mockWeb3(page)
+    const { switchChain } = await mockWeb3(page, {
+      accounts: [account],
+      chain: 'gnosis',
+    })
 
-    await page.goto('/new-route')
+    await page.goto(`/submit/${encode(route)}/${encode([transaction])}`)
 
-    await connectWallet(page)
-    await switchChain(10)
+    await connectWallet(page, account)
+    await switchChain(Chain.OETH)
 
     await expect(
-      page.getByRole('button', { name: 'Switch wallet to Ethereum' }),
+      page.getByRole('button', { name: 'Switch wallet to Gnosis' }),
     ).toBeInViewport()
   })
 })
